Authentication (#19)
Add authentication with phpunit tests

Also refactor ci workflow and add the following qa tools: stylelint and twigcs. They are included inside the makefile to simplify running commands.

About the template, the navbar, the header and the footer are created. An "alert" component is now available from the design system. Bootstrap 5 is also available for a faster development.

Add default fixtures, with LiipTestBundle and Dama Doctrine Bundle to enhance phpunit's features.

Update the docker-compose to add a persistent port : without it, phpunit is trying to use the localhost network and not the postgresql docker container.

Generate default translations for "fr" locale.